+++ projects/pip-webui2-themes/src/lib/shared/Theme.ts
@@ -1,6 +1,10 @@
+import { InjectionToken } from '@angular/core';
+
 export class Theme {
     name: string;
     palette = 'light';
 }
 
 export type Themes = Theme[];
+
+export const DEFAULT_THEME_NAME = new InjectionToken<string>('default theme name');