+++ projects/pip-webui2-themes/src/lib/shared/themes.service.ts
@@ -1,91 +1,35 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Inject } from '@angular/core';
 import { Observable, BehaviorSubject } from 'rxjs';
-import { Theme, Themes } from './Theme';
 
-export const defaultTheme = 'pip-blue-theme';
+import { Theme } from './Theme';
+import { THEMES_CONFIG, ThemesConfig } from './ThemeConfig';
 
 @Injectable()
 export class PipThemesService {
     private _selectedTheme: Theme = new Theme();
-    private _themes: Themes = [
-        {
-            name: 'candy-theme',
-            palette: 'light'
-        },
-        {
-            name: 'unicorn-dark-theme',
-            palette: 'dark'
-        },
-        {
-            name: 'pip-blue-theme',
-            palette: 'light'
-        },
-        {
-            name: 'pip-grey-theme',
-            palette: 'light'
-        },
-        {
-            name: 'pip-navy-theme',
-            palette: 'light'
-        },
-        {
-            name: 'pip-amber-theme',
-            palette: 'light'
-        },
-        {
-            name: 'pip-green-theme',
-            palette: 'light'
-        },
-        {
-            name: 'pip-orange-theme',
-            palette: 'light'
-        },
-        {
-            name: 'pip-pink-theme',
-            palette: 'light'
-        },
-        {
-            name: 'pip-dark-theme',
-            palette: 'dark'
-        },
-        {
-            name: 'pip-black-theme',
-            palette: 'dark'
-        },
-        {
-            name: 'bootbarn-warm-theme',
-            palette: 'light'
-        },
-        {
-            name: 'bootbarn-cool-theme',
-            palette: 'light'
-        },
-        {
-            name: 'bootbarn-mono-theme',
-            palette: 'light'
-        }
-    ];
 
     private _themes$;
     private _selectedTheme$;
     private _selectedThemeName$;
 
-    public constructor() {
+    public constructor(
+        @Inject(THEMES_CONFIG) private config: ThemesConfig
+    ) {
         // Load theme name from the local storage
         const name: string = window.localStorage.getItem('theme');
 
         // Set theme name
-        for (const theme of this._themes) {
+        for (const theme of this.config.themes) {
             if (theme.name === name) {
                 this._selectedTheme = theme;
                 break;
-            } else if (this._selectedTheme == null || theme.name === defaultTheme) {
+            } else if (this._selectedTheme == null || theme.name === this.config.defaultThemeName) {
                 this._selectedTheme = theme;
             }
         }
 
         // Create observables
-        this._themes$ = new BehaviorSubject<Theme[]>(this._themes);
+        this._themes$ = new BehaviorSubject<Theme[]>(this.config.themes);
         this._selectedTheme$ = new BehaviorSubject<Theme>(this._selectedTheme);
         this._selectedThemeName$ = new BehaviorSubject<string>(this._selectedTheme.name);
         this.selectedTheme = this._selectedTheme;
@@ -96,11 +40,11 @@ export class PipThemesService {
     }
 
     public get themes(): Theme[] {
-        return this._themes;
+        return this.config.themes;
     }
 
     public set themes(themes: Theme[]) {
-        this._themes = themes;
+        this.config.themes = themes;
         this._themes$.next(themes);
     }
 
@@ -109,7 +53,7 @@ export class PipThemesService {
     }
 
     public set selectedThemeName(name: string) {
-        for (const theme of this._themes) {
+        for (const theme of this.config.themes) {
             if (theme.name === name) {
                 this.selectedTheme = theme;
                 break;