+++ projects/pip-webui2-themes/src/lib/themes.module.ts
@@ -1,5 +1,6 @@
 import { NgModule, ModuleWithProviders } from '@angular/core';
 
+import { THEMES_CONFIG, DEFAULT_THEMES_CONFIG } from './shared/ThemeConfig';
 import { PipThemesService } from './shared/themes.service';
 
 @NgModule()
@@ -7,7 +8,13 @@ export class PipThemesModule {
   static forRoot(): ModuleWithProviders {
     return {
       ngModule: PipThemesModule,
-      providers: [PipThemesService]
+      providers: [
+        PipThemesService,
+        {
+          provide: THEMES_CONFIG,
+          useValue: DEFAULT_THEMES_CONFIG
+        }
+      ]
     };
   }
 }